back-end/yuccanleadfaq/urls.py
from django.urls import path, include
from authentication.views import SignUpView, LoginView, UsersList
from django.contrib import admin
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)


urlpatterns = [
    path('admin/', admin.site.urls),
    # path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    # path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/register/', SignUpView.as_view(), name='signup'),
    path('api/login/', LoginView.as_view(), name='login'),
    path('api/users/', UsersList.as_view(), name='users'),
    path('api/', include('questions.urls')),
]
