back-end/requirement.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
mysqlclient==2.1.1
pipenv==2023.2.4
platformdirs==3.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
virtualenv==20.19.0
virtualenv-clone==0.5.7
